allow for client-facing dataclasses to use any backend type
With services, the generated schema is important, and we require knowledge of implemented backends at runtime. With clients, the schema is not that important; they need to be able to specify a backend without having the backend installed
