+++ scripts/manual_client.py
@@ -24,7 +24,7 @@ from intersect_sdk import (
 from dial_dataclass import (
     DialInputPredictions,
     DialInputSingleOtherStrategy,
-    DialWorkflowCreationParams,
+    DialWorkflowCreationParamsClient,
     DialWorkflowDatasetUpdate,
 )
 
@@ -130,7 +130,7 @@ class ActiveLearningOrchestrator:
     # create a message to send to the server
     def assemble_message(self, operation: str, **kwargs: Any) -> IntersectClientCallback:
         if operation == 'initialize_workflow':
-            payload = DialWorkflowCreationParams(
+            payload = DialWorkflowCreationParamsClient(
                 dataset_x=INITIAL_DATASET_X,
                 dataset_y=INITIAL_DATASET_Y,
                 bounds=BOUNDS,
